Refactor password reset into one file
