--- src/components/account/PasswordReset/constants/defaultState.ts
@@ -1,5 +1,0 @@
-import { PaswordResetFormFieldsType } from "../types";
-
-export const defaultLoginFormFieldsState: PaswordResetFormFieldsType = {
-  email: "",
-};
