--- src/components/account/PasswordReset/hooks/usePasswordReset/types.ts
@@ -1,5 +1,0 @@
-import { PaswordResetFormFieldsType } from "../../types";
-
-export type PasswordResetActionsType = {
-  sendEmail: (data: PaswordResetFormFieldsType) => void;
-};
