--- src/components/account/PasswordReset/schema.ts
@@ -1,8 +1,0 @@
-import { errorMessages } from "../common/errorMessages";
-import { object, string } from "yup";
-
-export const passwordResetSchema = object()
-  .shape({
-    email: string().required().email(errorMessages.email),
-  })
-  .required();
