--- src/components/account/PasswordReset/types.ts
@@ -1,3 +1,0 @@
-export type PaswordResetFormFieldsType = {
-  email: string;
-};
