--- src/components/account/common/errorMessages.ts
@@ -1,5 +1,0 @@
-export const errorMessages = {
-  name: "Invalid e-mail address",
-  email: "Invalid e-mail address",
-  password: "Password is too short",
-};
